fix off by one in refactored resampler
