+++ src/mixer/tests.rs
@@ -79,6 +79,7 @@ fn rate_conv() {
     assert_eq!(rx.next(), Some(15.0));
     assert_eq!(rx.next(), Some(5.0));
     assert_eq!(rx.next(), Some(-5.0));
+    assert_eq!(rx.next(), Some(-2.5));
     assert_eq!(rx.next(), None);
 }
